+++ polls/views.py
@@ -1,14 +1,12 @@
-from rest_framework import status
-from rest_framework.decorators import action
 from rest_framework.generics import get_object_or_404
-from rest_framework.response import Response
-from rest_framework.viewsets import GenericViewSet
+from rest_framework.decorators import action
+from rest_framework.viewsets import GenericViewSet, ModelViewSet
 from rest_framework.mixins import ListModelMixin, RetrieveModelMixin, CreateModelMixin, \
     DestroyModelMixin
 
 from choose_one.paginators import ResultSetPagination
 from polls.serializers import PollSerializer, OptionSerializer
-from polls.models import Poll
+from polls.models import Poll, Option
 from polls.permissions import IsAuthenticatedOrReadOnly
 
 
@@ -33,6 +31,11 @@ class PollViewSet(GenericViewSet, ListModelMixin, RetrieveModelMixin, CreateMode
     Remove poll by its slug
 
     Remove poll by its slug, user should be creator of this poll
+
+    my_polls:
+    Return user's polls
+
+    Return user's polls
     """
     serializer_class = PollSerializer
     lookup_field = 'slug'
@@ -44,11 +47,54 @@ class PollViewSet(GenericViewSet, ListModelMixin, RetrieveModelMixin, CreateMode
             return Poll.objects.filter(creator=self.request.user)
         return Poll.objects.all()
 
-    @action(methods=['POST'], detail=True, url_name='add-option', url_path='add-option',
-            serializer_class=OptionSerializer)
-    def add_option(self, request, slug=None, *args, **kwargs):
-        poll = get_object_or_404(Poll, slug=slug, creator=self.request.user)
-        serializer = self.get_serializer_class()(data=request.data, context={'poll': poll})
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(status=status.HTTP_201_CREATED, data=serializer.data)
+    @action(detail=False, methods=['GET'], url_path='my-polls', url_name='my-polls')
+    def my_polls(self, request, *args, **kwargs):
+        queryset = self.get_queryset().filter(creator=self.request.user)
+        page = self.paginate_queryset(queryset)
+        serializer = self.get_serializer(page, many=True)
+        return self.get_paginated_response(serializer.data)
+
+
+class OptionViewSet(ModelViewSet):
+    """
+    list:
+    Return list of options
+
+    Return list of poll's options
+
+    retrieve:
+    Return single option
+
+    Return single poll's option
+
+    create:
+    Create new option
+
+    Create new poll's option
+
+    destroy:
+    Remove single option
+
+    Remove single poll's option
+
+    update:
+    Update single option
+
+    Update single poll's option
+
+    partial_update:
+    Partial Update single option
+
+    Partial Update single poll's option
+    """
+    serializer_class = OptionSerializer
+    permission_classes = (IsAuthenticatedOrReadOnly,)
+
+    def get_queryset(self):
+        return Option.objects.filter(poll__slug=self.kwargs['poll_slug'])
+
+    def get_serializer_context(self):
+        context = super().get_serializer_context()
+        poll = get_object_or_404(Poll, slug=self.kwargs['poll_slug'])
+        context['poll'] = poll
+        return context